Feature/webhook event dispatcher (#20)
* Start working on webhook event dispatcher

* Move processors to Core project

* Minor refactoring

* Create set webhook and set commands script

* Ignore bicep key warning

* Create github secret from pipeline

* Add mask to host key

* Turn webhook key into a secure string

* Retrieve function key from webhook script

* Use azure powershell script to execute the set webhook

* Remove debugging script messgaes

* Apply code factor suggestions

* Remove build-deploy.yaml trailling space

+++ src/MobileDataUsageReminder.Core/Infrastructure/Repository/Models/MobileData.cs
@@ -7,7 +7,7 @@ public class MobileData
     public int InitialAmount { get; init; }
     public int UsedAmount { get; init; }
     public int RemainingAmount { get; init; }
-    public int UsedPercentage { get; init; }
+    public double UsedPercentage { get; init; }
     public DateTime FullDate { get; init; }
     public int Day { get; init; }
     public string Month { get; init; }
